updated snapshots, refactor
