+++ src/app/containers/RadioSchedule/Canonical/index.jsx
@@ -14,12 +14,10 @@ import {
   GEL_SPACING_QUAD,
 } from '@bbc/gel-foundations/spacings';
 import RadioSchedule from '@bbc/psammead-radio-schedule';
-import moment from 'moment';
-import findLastIndex from 'ramda/src/findLastIndex';
-import propSatisfies from 'ramda/src/propSatisfies';
 import SectionLabel from '@bbc/psammead-section-label';
 import { Link } from '@bbc/psammead-story-promo';
 import { ServiceContext } from '#contexts/ServiceContext';
+import processRadioSchedule from '../utilities/processRadioSchedule';
 import webLogger from '#lib/logger.web';
 
 const logger = webLogger();
@@ -46,6 +44,10 @@ const RadioScheduleSection = styled.section.attrs(() => ({
 
 const FrontPageRadioScheduleSection = styled(RadioScheduleSection)`
   margin: 0 auto;
+  width: 100%; /* Needed for IE11 */
+  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) {
+    max-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN};
+  }
 `;
 
 const RadioScheduleSectionLabel = styled(SectionLabel)`
@@ -70,69 +72,11 @@ const CanonicalRadioSchedule = ({ endpoint }) => {
     radioSchedule: { header, frequenciesPageUrl, frequenciesPageLabel },
   } = useContext(ServiceContext);
 
-  const getProgramState = (currentTime, startTime, endTime) => {
-    const isLive = currentTime < endTime && currentTime > startTime;
-    if (isLive) {
-      return 'live';
-    }
-    const hasEnded = currentTime > endTime;
-    if (hasEnded) {
-      return 'onDemand';
-    }
-    return 'next';
-  };
-
   useEffect(() => {
-    const getLink = (state, program) => {
-      const url = `/${service}/${program.serviceId}`;
-      return state === 'live'
-        ? `${url}/liveradio`
-        : `${url}/${program.broadcast.pid}`;
-    };
-
     const handleResponse = async response => {
       const radioScheduleData = await response.json();
 
-      const currentTime = parseInt(moment.utc().format('x'), 10);
-      // finding latest program, that may or may not still be live. this is because there isn't
-      // always a live program, in which case we show the most recently played program on demand.
-      const latestProgramIndex = findLastIndex(
-        propSatisfies(time => time < currentTime, 'publishedTimeStart'),
-      )(radioScheduleData.schedules);
-
-      const radioSchedules = radioScheduleData.schedules;
-
-      const scheduleDataIsComplete =
-        radioSchedules[latestProgramIndex - 2] &&
-        radioSchedules[latestProgramIndex + 1];
-      const schedulesToShow = scheduleDataIsComplete && [
-        radioSchedules[latestProgramIndex],
-        radioSchedules[latestProgramIndex - 1],
-        radioSchedules[latestProgramIndex - 2],
-        radioSchedules[latestProgramIndex + 1],
-      ];
-
-      const schedules =
-        schedulesToShow &&
-        schedulesToShow.map(program => {
-          const currentState = getProgramState(
-            currentTime,
-            program.publishedTimeStart,
-            program.publishedTimeEnd,
-          );
-          return {
-            id: program.broadcast.pid,
-            state: currentState,
-            stateLabel: currentState,
-            startTime: program.publishedTimeStart,
-            link: getLink(currentState, program),
-            brandTitle: program.brand.title,
-            episodeTitle: program.episode.presentationTitle,
-            summary: program.episode.synopses.short,
-            duration: program.publishedTimeDuration,
-            durationLabel: 'Duration',
-          };
-        });
+      const schedules = processRadioSchedule(radioScheduleData, service);
       setRadioSchedule(schedules);
     };
